requirements.txt
black==19.10b0
pre-commit==2.6.0
pytest-homeassistant-custom-component>=0.4.4
paho-mqtt>=1.6.1
websocket-client>=1.2.3